+++ CHANGELOG.md
@@ -367,3 +367,7 @@
 
 
 (TTHERE IS A BIG PROBLEM, THE PIECES GET CROSSED EACH OTHER) this not ended yet.
+
+##[2024-10-22]
+## Added:
+- Added a new mwthof in `Board` `path_is_clear`  to check if a path is clear for a piece to move to a new position (i hope to  fix the problem of pieces crossing each other).